delete test authorization fix

+++ Backend/studycave-rest-service/src/main/java/studycave/application/test/TestController.java
@@ -12,6 +12,8 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,6 +21,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -173,8 +176,21 @@ public class TestController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void deleteTest(@PathVariable(required = true) Long id) {
-		testRepository.deleteById(id);
+	public ResponseEntity deleteTest(@RequestHeader(value = "Authorization") String headerStr,
+			@PathVariable(required = true) Long id) {
+
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String currentPrincipalName = authentication.getName();
+		Long userId = userRepository.findByUsername(currentPrincipalName).get().getId();
+
+		Optional<Test> test = testRepository.findById(id);
+		if (test.isPresent()) {
+			if (userId.equals(test.get().getIdOwner())) {
+				testRepository.deleteById(id);
+				return new ResponseEntity(HttpStatus.OK);
+			}
+		}
+		return new ResponseEntity("Access Forbidden",HttpStatus.FORBIDDEN);
 	}
 
 	@PostMapping
